pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus"
version = "0.1.0"
description = "NEXUS — The Institutional Memory Engine"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Oussama Iskia (SKAY)", email = "skay@nexus.dev" },
]
dependencies = [
    # Web
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "python-multipart==0.0.9",
    # RAG / LangChain (LCEL only, no legacy chains)
    "langchain>=0.3.0,<0.4",
    "langchain-community>=0.3.0,<0.4",
    "langchain-groq>=0.2.0,<0.3",
    "langchain-ollama>=0.2.0,<0.3",
    "langchain-text-splitters>=0.3.0,<0.4",
    # Document parsing
    "unstructured>=0.16,<0.19",
    "pypdf>=4.0,<5",
    "python-docx>=1.1,<2",
    # Embeddings
    "sentence-transformers==3.0.1",
    # Keyword search
    "rank-bm25==0.2.2",
    # Reranker
    "transformers>=4.40,<5",
    # Vector DB (Supabase pgvector)
    "supabase>=2.7,<3",
    # LLM clients
    "groq>=0.11,<1",
    # Logging
    "structlog>=24.1,<25",
    # Config
    "pydantic>=2.8,<3",
    "pydantic-settings>=2.4,<3",
    # Utils
    "httpx>=0.27,<1",
    "aiofiles>=24.1,<25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5.0,<6",
    "ruff>=0.6,<1",
    "black>=24.0,<25",
    "mypy>=1.11,<2",
    "pre-commit>=3.8,<4",
    "types-aiofiles>=24.1,<25",
]
eval = [
    "ragas>=0.2.0,<0.3",
    "datasets>=2.21,<3",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # type-checking imports
]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["nexus"]

[tool.black]
target-version = ["py310"]
line-length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "nexus.features.*"
strict = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"
markers = [
    "integration: marks tests requiring external services (deselect with '-m \"not integration\"')",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nexus"]
